Initial move of packages to dedicated repo
The wiki _could_ be modified by anyone and could cause confusion. To
avoid this, I propose we move the packages to a dedicated repository
that takes PRs to ensure accuracy.

This is by no means the way we should do it, but more so I am looking
for validation of this idea.